Explicitly convert call to sqrt() to an unsigned int

+++ src/demos/many-cubes.cpp
@@ -65,7 +65,7 @@ namespace manyCubesDemo {
     ammonite::renderer::drawFrame();
 
     //Reposition cubes
-    sideLength = sqrt(modelCount);
+    sideLength = (unsigned int)sqrt(modelCount);
     for (unsigned int x = 0; x < sideLength; x++) {
       for (unsigned int y = 0; y < sideLength; y++) {
         ammonite::models::position::setPosition(loadedModelIds[((std::size_t)x * sideLength) + y],
